修复一处size_t和int转换的bug。优化demo内容

+++ algorithmCtrl/hnsw/hnswAlgo/hnswalg.h
@@ -741,7 +741,7 @@ namespace hnswlib {
           label_c = search->second;
 
           char* data_ptrv = getDataByInternalId(label_c);
-          size_t dim = *((size_t *) dist_func_param_);
+          int dim = (int)*((size_t *) dist_func_param_);    // 不转成int，会出偶现的问题
           std::vector<data_t> data;
           data_t* data_ptr = (data_t*) data_ptrv;
           for (int i = 0; i < dim; i++) {
